requirements.txt
alembic==1.0.7
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.12.1
cryptography==2.5
dataset==1.1.2
emoji==0.5.1
future==0.17.1
Mako==1.0.7
MarkupSafe==1.1.0
pycparser==2.19
python-dateutil==2.8.0
python-editor==1.0.4
python-telegram-bot==11.1.0
six==1.12.0
SQLAlchemy==1.2.18
